% Auto-generated: do not edit by hand
\name{''column}

\alias{''column}

\title{column component}

\description{

}

\usage{
''column(column=NULL, items=NULL, showHandle=NULL)
}

\arguments{
\item{column}{Named list. }

\item{items}{Unnamed list. The ID used to identify this component in Dash callbacks.}

\item{showHandle}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

