% Auto-generated: do not edit by hand
\name{''DashDraggable}

\alias{''DashDraggable}

\title{DashDraggable component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''DashDraggable(id=NULL, columnOrder=NULL, columns=NULL, handleText=NULL,
items=NULL, showHandle=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{columnOrder}{Unnamed list. }

\item{columns}{Named list. }

\item{handleText}{Character. }

\item{items}{Named list. }

\item{showHandle}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

