% Auto-generated: do not edit by hand
\name{''item}

\alias{''item}

\title{item component}

\description{

}

\usage{
''item(index=NULL, item=NULL)
}

\arguments{
\item{index}{Numeric. }

\item{item}{Named list. The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

