% Auto-generated: do not edit by hand
\name{''task}

\alias{''task}

\title{task component}

\description{

}

\usage{
''task(index=NULL, task=NULL)
}

\arguments{
\item{index}{Numeric. }

\item{task}{Named list. The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

