% Auto-generated: do not edit by hand
\name{''column}

\alias{''column}

\title{column component}

\description{

}

\usage{
''column(column=NULL, tasks=NULL)
}

\arguments{
\item{column}{Named list. }

\item{tasks}{Unnamed list. The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

