% Auto-generated: do not edit by hand
\name{''item}

\alias{''item}

\title{item component}

\description{

}

\usage{
''item(handleText=NULL, index=NULL, item=NULL, showHandle=NULL)
}

\arguments{
\item{handleText}{Character. }

\item{index}{Numeric. }

\item{item}{Named list. The ID used to identify this component in Dash callbacks.}

\item{showHandle}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

